HW-1 Change projects structure: do not create separate project for every task
