--- lesson-three/.idea/eclipseCodeFormatter.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="EclipseCodeFormatterProjectSettings">
-    <option name="projectSpecificProfile">
-      <ProjectSpecificProfile>
-        <option name="configType" value="ECLIPSE" />
-        <option name="formatter" value="ECLIPSE" />
-      </ProjectSpecificProfile>
-    </option>
-  </component>
-</project>
