--- lesson-three/src/main/java/ru/mtsb/okovalev/java/learning/lessonthree/Main.java
@@ -1,26 +1,0 @@
-package ru.mtsb.okovalev.java.learning.lessonthree;
-
-import java.util.ArrayList;
-
-import ru.mtsb.okovalev.java.learning.lessonthree.animals.Animal;
-
-public class Main {
-	@SuppressWarnings("all")
-	public static void main(String[] args) {
-		CreateAnimalsServiceImpl createAnimalsServiceImpl = new CreateAnimalsServiceImpl();
-		ArrayList<Animal> byCreateAnimalsServiceCreate = createAnimalsServiceImpl.createDefault();
-		for (int i = 0; i < byCreateAnimalsServiceCreate.size(); i++) {
-			System.out.println(byCreateAnimalsServiceCreate.get(i));
-		}
-
-		System.out.println();
-		ArrayList<Animal> byCreateAnimalsServiceImplCreate = createAnimalsServiceImpl.create();
-		for (Animal animal : byCreateAnimalsServiceImplCreate) {
-			System.out.println(animal);
-		}
-
-		System.out.println();
-		ArrayList<Animal> byCreateAnimalsServiceImplCreateN = createAnimalsServiceImpl.create(12);
-		byCreateAnimalsServiceImplCreateN.forEach(System.out::println);
-	}
-}
