--- lesson-three/src/main/java/ru/mtsb/okovalev/java/learning/lessonthree/animals/Animal.java
@@ -1,11 +1,0 @@
-package ru.mtsb.okovalev.java.learning.lessonthree.animals;
-
-public interface Animal {
-	void move();
-
-	void eat();
-
-	void sound();
-
-	void live();
-}
