--- lesson-two/gradle.properties
@@ -1,5 +1,0 @@
-# all - shows detailed information about all the problems in *.gradle file
-# fail - fails the Gradle task if any of problems has been found in *.gradle file
-# summary - shows short description of warning
-# none - doesn't show any warnings
-org.gradle.warning.mode=none
