+++ src/main/java/ru/mtsb/okovalev/lessonthree/animals/Animal.java
@@ -1,0 +1,11 @@
+package ru.mtsb.okovalev.lessonthree.animals;
+
+public interface Animal {
+	void move();
+
+	void eat();
+
+	void sound();
+
+	void live();
+}
